No longer force passing actor instance to enable/disable.
This is related to #13. First, it allows just passing in anything that
responds to id to enable/disable and friends. Second, it wraps whatever
is passed in, if possible, and uses the wrapped value. Third,
identifier is no longer supported. From now on id.to_s will be used
everywhere.