Dramatically simplify life by forcing flipper_id instead of id.
The big problem with just checking id is that EVERYTHING responds to id
even if it doesn't, due to the ruby deprecation warning for using
object_id instead of id.

#1. This change makes it super easy to detect actors because no one
will have a flipper_id method already in their models.

#2. This allows people to change what is used for activating individual
actors. If you don't want it to be id, simply define flipper_id to be
name, or whatever else you want.