Conditional gates should not be memoized
It is based on gate_values which changes outside of the memoizing process. This means it either needs to memoize based on gate values or not memoize at all. I believe not memoizing at all is better in this instance since the method is only used in state and description, which are mostly for UI things, not hot code paths.