Use the instrumented block's payload, not the original payload
Some instrumentation services don't pass the original payload straight
through to the block (e.g., they might copy it and add their own
information to it). With those instrumenters it's important to use the
payload passed to the block rather than the original payload or
information added by the block might get lost.