Allow an extra parameter to be passed to group blocks
Methods that took a group name now take an optional block parameter,
which is passed through to the group's block. This can be used, e.g., to
implement dynamic groups based on a database. Here's how it looks:

    Flipper.register(:team) do |actor, team_id|
      Team.find(team_id).members.include?(actor)
    end

    $flipper[:my_feature].enable $flipper.group(:team, Team.by_name("admins").id.to_s)

The block parameter must be convertible to a String via #to_str.