Ensure that memoizable, instrumented and operation logger delegate
When you have custom adapters that wrap others and expect those methods to be present, it can be confusing when they aren't because flipper wraps your adapter passed in with instrumented and even memoizable if you use that. For example, we have some custom ones at github with other methods that get called in middleware and not having memoizable and instrumented delegate the methods to the original adapter passed in left me with no method errors when upgrading to 0.8.