Stop instrumenting adapters by default
Before it was assumed you would want to instrument adapter operations always and there was no way to back out of that. Now, instrumenting adapters is opt in. Do it if you want. Leave it alone if you don't. If you are using custom adapters, you can track stats inside them however you want. No need to also get this instrumentation. This allows for the most flexibility.