Allows features to be retrieve in bulk
When N features are being checked in a single request this would result
in N requests to the underlying store at best with memoization enabled.

This commit adds `get_multi` to the adapter protocol, backfilled to all
adapters through the Flipper::Adapter module implementing it as multiple
`get` calls.

This allows adapters to optimize `get_multi` when possible to improve
the performance of retrieving multiple flags.

To take advantage of this `preload` and `preload_all` were added to the
DSL, which allow the specified features to eagerly memoized. A new eager
version of the memoizing middleware was added to demonstrate this.

`get_multi` is currently only implement for the Memoizable and Redis
adapters.