Make rollout percentage equal to 100 equal to flipper fully enabled
Rollout doesn't remove other gate values when fully activated. Rollout
at 100% is equal to flipper fully enabled in that rollout returns true
for active? when percentage is 100. Flipper only returns true for 100%
percentage of actors if an actor is passed. If no actor is passed, it
returns false unless fully enabled or % of time is true.