Allow configuring a local adapter for cloud
This improves latency and resiliency. Latency is improved because cloud
now uses a sync adapter with reads going to the local. Local reads mean
low latency.

Resiliency is improved because reads go to local which means if cloud is
unavailable most reads will be fine other than a few slow ones when
syncs are attempted (because they are not in background thread
currently).