Ensure that locally enabled features properly sync conditional enablements
The scenario that was failing was if the local was enabled, then the
feature was disabled remotely and a single actor or other conditional
enablement was modified, the local enabled feature would ignore the
update.

A feature cannot be fully enabled and conditionally enabled. Because we
don't pass the entire log of changes and instead just do a diff, we need
to make sure that a local feature is disabled prior to syncing
conditional enablements.