Allow synchronizer to raise selectively
This makes it so it can raise on import, which you want feedback on, but
not raise on sync, which would interrupt whatever was going on in a
request or job.

Might want to change this to Synchronizer (raises) and
InstrumentedSynchronizer (exceptions rescued and instrumented) or
something to indicate the difference and avoid a silly option.