Add support for slashes in feature names to UI
refs #357

This is a tricky one. The issue is that I was assuming no one would use slashes. Using a slash completely broke all the routing. I changed the routing to be a block so it is more flexible. Now it is possible to route based on anything in request.

The API will likely need the same work.