Unescapes feature_name in FeatureNameFromRoute
* Before we supported slashes we unescaped the feature name, but teh
change to pull feature_name using  FeatureNameFromRoute is not
unescaping feature names.  This is an issue because when you create a
new feature it redirects to the feature show page.  Then when you go to
enable a gate the feature_key the gate gets is not unescaped so the
feature is not enabled.