Allow shorthand block notation on group types
When a shorthand block was used when initializing a
Group, an error would be thrown about not passing
the correct number of arguments. That was due to
the arity of the shortand being -1 instead of 1.
By taking the absolute value of the arity, an arity
of -1 will still correctly set the single_argument
instance variable.