Remove descriptions from api
It brings up a larger question of should the descriptions source be in Flipper::UI or Flipper::API or both or moved to Flipper config itself so it can be re-used and I'd rather see what people have to say.

Remove as_json because it isn't even used in flipper ui and api has its own.