100% of actors now considered conditional rollout
The tldr is that 100% of actors could still return false if the actor doesn't include traffic from users that are not signed in. It would be true for 100% of sessions where session was the actor but not for 100% of users where users may be signed in or signed out. Signed out user would be nil and would return false for enabled.