fix(sequel): change get_all sql to sequel DSL
I can't figure out how to get the joined sequel result class
to respond to all the selected columns as attibutes. It seems
that if you use the feature model to query then the object
that comes back is a feature model not a composite of features
and gates. Since the original query actually queried using the
gates class I've taken the features sql and queried using the
gates model so as not to affect the original downstream implementation
that expects the result to respond to messages implemented by the
gate class. At some point this could be refactored to be more
succinct.