Remove use of Rack::BodyProxy
Rack::BodyProxy allows deferring execution until after the response has sent to the client. It's intended for expensive cleanup that could delay delivering the response. Cleaning up memoization is not an expensive operation, and I suspect this use of BodyProxy is the source for warnings in https://github.com/jnunemaker/flipper/issues/604