Remove AS.on_load hook for ActiveRecord adapter
cc https://github.com/jnunemaker/flipper/issues/634

This load hook causes the order of ActiveRecord loading to be indeterminte. In a pristine Rails app, AR does not load until after all the initializers are run, which overwrites the configured Flipper adapter in an initializer. This changes make the AR adapter consistent with the others that configure themselves (redis, mongo, etc).