Decouple confirmation text in separate UI buttons
I noticed that the confirmation text set up by `confirm_fully_enable` does not work correctly if `feature_removal_enabled` was set to `false`.

This was due to the node with ID `feature_name` only being present within the form included when `feature_removal_enabled` was set to true. This lead to a situation where your UI would display "undefined" as the prompt text though when that or the actual feature name was used the  flag was not toggled since it was comparing the string "undefined" to an undefined variable.

This commit fixes this edge case by decoupling the confirmation text for both buttons from the `feature_removal_enabled` form and moving them into data attributes on the button being clicked. This allows them to be enabled/disabled separately without impacting each other. The appearance of these attributes are controlled by the flags themselves.