Improve http adapter to do single http request for import
Before it was N where N was the number of differences between http adapter and source adapter.

Also, simple delegator causes weirdness. Methods in Flipper::Adapter run instead of delegating to the underlying adapter. Removed that in favor of explicitness.

Renamed input to contents because that seemed more appropriate.

Cleaned up subscriptions that sometimes were getting setup and then continuing to be used for later specs.

Noop'd a couple instrumentation things that were outputting warnings.