Check existence of RSpec.configure
When Rails commands such as db:migrate are executed (formerly rake
tasks), they load the broader set of rake tasks, and that means that
RSpec's rake task is loaded without all of the RSpec context - which
means there _is_ an RSpec constant, but it does not have the configure
method.
i.e. lib/rspec/core/rake_task.rb in rspec-core

I considered loading `rspec/core` here instead, but that feels counter
to what's needed: if we have that full RSpec context, then we should
add the Flipper helpers. But if we don't, then we should respect the
current context, not load all of RSpec, and not add Flipper's helpers.