More sleep time between polls
Before we were trying to ensure it was interval or as fast as the request took. This is an issue if we are struggling to serve requests as it increases throughput when requests are slow instead of slowing them down.

Need to probably back off even more if response time is more than 10 seconds but can look at that later.